Fix[#311] : AppDelegate 와 RootViewController 인 ViewController 에서 두 번 날씨 정보를 요청하고 있습니다.
- Archive 시 Upload 가 되지 않는 이슈가 있습니다.
