+++ Kloudy/Kloudy/Application/AppDelegate.swift
@@ -12,10 +12,8 @@ import RxSwift
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate, CLLocationManagerDelegate {
-    var locationCount = CoreDataManager.shared.countLocations()
-    
     let dummyData = FetchWeatherInformation().dummyData
-    lazy var weathers = [Weather](repeating: dummyData , count: locationCount+1)
+    var weathers = [Weather]()
     let disposeBag = DisposeBag()
     lazy var coreDataStack: CoreDataStack = .init(modelName: "Kloudy")
     
@@ -28,42 +26,96 @@ class AppDelegate: UIResponder, UIApplicationDelegate, CLLocationManagerDelegate
     
     // 어플리케이션의 런치 프로세스가 끝났을 때 -> Fetch 요청을 보냄, 요청이 끝나고 난 후 메인화면으로 넘어감.
     func application(_ application: UIApplication, willFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
-        let currentStatus = CLLocationManager().authorizationStatus
-        if locationCount == 0 {
-            if (currentStatus == .restricted || currentStatus == .notDetermined || currentStatus == .denied) {
-                self.weathers[0] = dummyData
-                CityWeatherNetwork().fetchCityWeather(code: "1111000000")
-                    .subscribe { event in
-                        switch event {
-                        case .success(let data):
-                            self.weathers.append(data)
-                        case .failure(let error):
-                            print("Error: ", error)
-                        }
-                    }
-                    .disposed(by: disposeBag)
-                CoreDataManager.shared.saveLocation(code: "1111000000", city: "Jongno-gu", province: "Seoul", sequence: CoreDataManager.shared.countLocations(), indexArray: ["rain", "mask", "laundry", "car", "outer"])
-            } else {
-                let XY = LocationManager.shared.requestNowLocationInfo()
-                let nowLocation = FetchWeatherInformation.shared.getLocationInfoByXY(x: XY[0], y: XY[1])
-                guard let nowLocation = nowLocation else { return true }
-                
-                CityWeatherNetwork().fetchCityWeather(code: nowLocation.city)
-                    .subscribe { event in
-                        switch event {
-                        case .success(let data):
-                            self.weathers[0] = data
-                        case .failure(let error):
-                            print("Error: ", error)
-                        }
-                    }
-                    .disposed(by: disposeBag)
-            }
-        }
-        else {
-            self.weathers[0] = dummyData
-            loadLocation()
-        }
+//        let currentStatus = CLLocationManager().authorizationStatus
+//        
+//        let locations = CoreDataManager.shared.fetchLocations()
+//        weathers = [Weather](repeating: dummyData , count: locations.count + 1)
+//        print(locations)
+//        
+//        if locations.count == 0 {
+//            if (currentStatus == .restricted || currentStatus == .notDetermined || currentStatus == .denied) {
+//                CityWeatherNetwork().fetchCityWeather(code: "1111000000")
+//                    .subscribe { event in
+//                        switch event {
+//                        case .success(let data):
+//                            self.weathers.append(data)
+//                        case .failure(let error):
+//                            print("Error: ", error)
+//                        }
+//                    }
+//                    .disposed(by: disposeBag)
+//                CoreDataManager.shared.saveLocation(code: "1111000000", city: "Jongno-gu", province: "Seoul", sequence: CoreDataManager.shared.countLocations(), indexArray: ["rain", "mask", "laundry", "car", "outer"])
+//            } else {
+//                let XY = LocationManager.shared.requestNowLocationInfo()
+//                let nowLocation = FetchWeatherInformation.shared.getLocationInfoByXY(x: XY[0], y: XY[1])
+//                guard let nowLocation = nowLocation else { return true }
+//                
+//                CityWeatherNetwork().fetchCityWeather(code: nowLocation.code)
+//                    .subscribe { event in
+//                        switch event {
+//                        case .success(let data):
+//                            self.weathers[0] = data
+//                        case .failure(let error):
+//                            print("Error: ", error)
+//                        }
+//                    }
+//                    .disposed(by: disposeBag)
+//            }
+//        }
+//        else {
+////            loadLocation()
+//            if (currentStatus == .restricted || currentStatus == .notDetermined || currentStatus == .denied) {
+//                for index in locations.indices {
+//                    CityWeatherNetwork().fetchCityWeather(code: locations[index].code ?? "")
+//                        .subscribe(on: ConcurrentDispatchQueueScheduler(qos: .default))
+//                        .subscribe { event in
+//                            switch event {
+//                            case .success(let data):
+//                                DispatchQueue.main.async {
+//                                    self.weathers[index + 1] = data
+//                                }
+//                            case .failure(let error):
+//                                print("Error: ", error)
+//                            }
+//                        }
+//                        .disposed(by: disposeBag)
+//                }
+//            } else {
+//                let XY = LocationManager.shared.requestNowLocationInfo()
+//                let nowLocation = FetchWeatherInformation.shared.getLocationInfoByXY(x: XY[0], y: XY[1])
+//                guard let nowLocation = nowLocation else { return true }
+//                
+//                CityWeatherNetwork().fetchCityWeather(code: nowLocation.code)
+//                    .subscribe(on: ConcurrentDispatchQueueScheduler(qos: .default))
+//                    .subscribe { event in
+//                        switch event {
+//                        case .success(let data):
+//                            DispatchQueue.main.async {
+//                                self.weathers[0] = data
+//                            }
+//                        case .failure(let error):
+//                            print("Error: ", error)
+//                        }
+//                    }
+//                    .disposed(by: disposeBag)
+//                
+//                for index in locations.indices {
+//                    CityWeatherNetwork().fetchCityWeather(code: locations[index].code ?? "")
+//                        .subscribe(on: ConcurrentDispatchQueueScheduler(qos: .default))
+//                        .subscribe { event in
+//                            switch event {
+//                            case .success(let data):
+//                                DispatchQueue.main.async {
+//                                    self.weathers[index + 1] = data
+//                                }
+//                            case .failure(let error):
+//                                print("Error: ", error)
+//                            }
+//                        }
+//                        .disposed(by: disposeBag)
+//                }
+//            }
+//        }
         return true
     }
     
@@ -87,7 +139,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate, CLLocationManagerDelegate
     }
     
     func loadLocation() {
-        lazy var myLocations = CoreDataManager.shared.fetchLocations()
+        let myLocations = CoreDataManager.shared.fetchLocations()
+//        for
         for locationIndex in myLocations.indices {
             // 지역 값이 뭔가 잘못된 것이 들어왔다면 끝내야함
             guard let province = myLocations[locationIndex].province else { return }
@@ -95,6 +148,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate, CLLocationManagerDelegate
             FetchWeatherInformation.shared.startLoad(province:province, city: city) { response in
                 self.weathers[locationIndex + 1] = response
             }
+            
+//            CityWeatherNetwork().fetchCityWeather(code: myLocations[locationIndex].code ?? "")
+            
         let currentStatus = CLLocationManager().authorizationStatus
             if currentStatus == .authorizedWhenInUse || currentStatus == .authorizedAlways {
                 // 현재 지역의 위도 경도로 기상청에서 제공하는 XY값을 계산 -> XY값으로 현재 지역정보 반환 후 요청을 보냄.
