+++ Kloudy/Kloudy/Application/ViewController.swift
@@ -8,6 +8,7 @@
 import Foundation
 import UIKit
 import CoreLocation
+import RxSwift
 
 class ViewController: UIViewController {
     
@@ -50,6 +51,10 @@ class ViewController: UIViewController {
         return logoImageView
     }()
     
+    let dummyData = FetchWeatherInformation().dummyData
+    var weathers = [Weather]()
+    let disposeBag = DisposeBag()
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         self.view.backgroundColor = UIColor.KColor.primaryBlue01
@@ -59,6 +64,7 @@ class ViewController: UIViewController {
         self.fetchWeatherDatas()
         DispatchQueue.main.asyncAfter(deadline: DispatchTime.now() + 3) {
             let checkWeatherView = CheckWeatherView()
+            checkWeatherView.weathers = self.weathers
             self.navigationController?.pushViewController(checkWeatherView, animated: false)
         }
     }
@@ -112,31 +118,120 @@ class ViewController: UIViewController {
     
     func fetchWeatherDatas() {
         
-        let cityInformationModel = FetchWeatherInformation()
-        lazy var cityData = cityInformationModel.loadCityListFromCSV()
-        let appDelegate = UIApplication.shared.delegate as? AppDelegate
-        
-        let myLocations = CoreDataManager.shared.fetchLocations()
-        for locationIndex in myLocations.indices {
-            // 지역 값이 뭔가 잘못된 것이 들어왔다면 끝내야함
-            guard let province = myLocations[locationIndex].province else { return }
-            guard let city = myLocations[locationIndex].city else { return }
-            FetchWeatherInformation.shared.startLoad(province:province, city: city) { response in
-                appDelegate?.weathers[locationIndex] = response
-            }
-        }
-        
+//        let cityInformationModel = FetchWeatherInformation()
+//        lazy var cityData = cityInformationModel.loadCityListFromCSV()
+//        let appDelegate = UIApplication.shared.delegate as? AppDelegate
+//
+//        let myLocations = CoreDataManager.shared.fetchLocations()
+//        for locationIndex in myLocations.indices {
+//            // 지역 값이 뭔가 잘못된 것이 들어왔다면 끝내야함
+//            guard let province = myLocations[locationIndex].province else { return }
+//            guard let city = myLocations[locationIndex].city else { return }
+//            FetchWeatherInformation.shared.startLoad(province:province, city: city) { response in
+//                appDelegate?.weathers[locationIndex] = response
+//            }
+//        }
+//
+//        let currentStatus = CLLocationManager().authorizationStatus
+//
+//        if currentStatus == .authorizedWhenInUse || currentStatus == .authorizedAlways {
+//            // 현재 지역의 위도 경도로 기상청에서 제공하는 XY값을 계산 -> XY값으로 현재 지역정보 반환 후 요청을 보냄.
+//            let XY = LocationManager.shared.requestNowLocationInfo()
+//            let nowLocation = FetchWeatherInformation.shared.getLocationInfoByXY(x: XY[0], y: XY[1])
+//            guard let nowLocation = nowLocation else { return }
+//            FetchWeatherInformation.shared.startLoad(province: nowLocation.province, city: nowLocation.city) { response in
+//                appDelegate?.weathers.append(response)
+//            }
+//        }
+//
         let currentStatus = CLLocationManager().authorizationStatus
-
-        if currentStatus == .authorizedWhenInUse || currentStatus == .authorizedAlways {
-            // 현재 지역의 위도 경도로 기상청에서 제공하는 XY값을 계산 -> XY값으로 현재 지역정보 반환 후 요청을 보냄.
-            let XY = LocationManager.shared.requestNowLocationInfo()
-            let nowLocation = FetchWeatherInformation.shared.getLocationInfoByXY(x: XY[0], y: XY[1])
-            guard let nowLocation = nowLocation else { return }
-            FetchWeatherInformation.shared.startLoad(province: nowLocation.province, city: nowLocation.city) { response in
-                appDelegate?.weathers.append(response)
+        
+        let locations = CoreDataManager.shared.fetchLocations()
+        weathers = [Weather](repeating: dummyData , count: locations.count + 1)
+        print(locations)
+        
+        if locations.count == 0 {
+            if (currentStatus == .restricted || currentStatus == .notDetermined || currentStatus == .denied) {
+                CityWeatherNetwork().fetchCityWeather(code: "1111000000")
+                    .subscribe { event in
+                        switch event {
+                        case .success(let data):
+                            self.weathers.append(data)
+                        case .failure(let error):
+                            print("Error: ", error)
+                        }
+                    }
+                    .disposed(by: disposeBag)
+                CoreDataManager.shared.saveLocation(code: "1111000000", city: "Jongno-gu", province: "Seoul", sequence: CoreDataManager.shared.countLocations(), indexArray: ["rain", "mask", "laundry", "car", "outer"])
+            } else {
+                let XY = LocationManager.shared.requestNowLocationInfo()
+                let nowLocation = FetchWeatherInformation.shared.getLocationInfoByXY(x: XY[0], y: XY[1])
+                guard let nowLocation = nowLocation else { return }
+                
+                CityWeatherNetwork().fetchCityWeather(code: nowLocation.code)
+                    .subscribe { event in
+                        switch event {
+                        case .success(let data):
+                            self.weathers[0] = data
+                        case .failure(let error):
+                            print("Error: ", error)
+                        }
+                    }
+                    .disposed(by: disposeBag)
             }
         }
-        
+        else {
+            if (currentStatus == .restricted || currentStatus == .notDetermined || currentStatus == .denied) {
+                for index in locations.indices {
+                    CityWeatherNetwork().fetchCityWeather(code: locations[index].code ?? "")
+                        .subscribe(on: ConcurrentDispatchQueueScheduler(qos: .default))
+                        .subscribe { event in
+                            switch event {
+                            case .success(let data):
+                                DispatchQueue.main.async {
+                                    self.weathers[index + 1] = data
+                                }
+                            case .failure(let error):
+                                print("Error: ", error)
+                            }
+                        }
+                        .disposed(by: disposeBag)
+                }
+            } else {
+                let XY = LocationManager.shared.requestNowLocationInfo()
+                let nowLocation = FetchWeatherInformation.shared.getLocationInfoByXY(x: XY[0], y: XY[1])
+                guard let nowLocation = nowLocation else { return }
+                
+                CityWeatherNetwork().fetchCityWeather(code: nowLocation.code)
+                    .subscribe(on: ConcurrentDispatchQueueScheduler(qos: .default))
+                    .subscribe { event in
+                        switch event {
+                        case .success(let data):
+                            DispatchQueue.main.async {
+                                self.weathers[0] = data
+                            }
+                        case .failure(let error):
+                            print("Error: ", error)
+                        }
+                    }
+                    .disposed(by: disposeBag)
+                
+                for index in locations.indices {
+                    CityWeatherNetwork().fetchCityWeather(code: locations[index].code ?? "")
+                        .subscribe(on: ConcurrentDispatchQueueScheduler(qos: .default))
+                        .subscribe { event in
+                            switch event {
+                            case .success(let data):
+                                DispatchQueue.main.async {
+                                    self.weathers[index + 1] = data
+                                }
+                            case .failure(let error):
+                                print("Error: ", error)
+                            }
+                        }
+                        .disposed(by: disposeBag)
+                }
+            }
+        }
     }
 }
