+++ Kloudy/Kloudy/View/CheckWeather/WeatherIndexView/LocationWeatherIndexView/LocationWeatherIndexView.swift
@@ -533,21 +533,21 @@ class LocationWeatherIndexView: UIView {
 extension LocationWeatherIndexView:  UICollectionViewDelegate, UICollectionViewDataSource,  UICollectionViewDelegateFlowLayout {
 
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        let indexName = self.indexArray[self.internalIndex]
-        let cellCount = calculateInternalIndexCount(indexName: indexName).count
-        return cellCount
+//        let indexName = self.indexArray[self.internalIndex]
+//        let cellCount = calculateInternalIndexCount(indexName: indexName).count
+        return 0
     }
 
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: InternalIndexCollectionViewCell.identifier, for: indexPath)
-        let indexName = self.indexArray[self.internalIndex]
-        let isIndexOn = calculateInternalIndexCount(indexName: indexName)
-        let internalIndexView = findInternalIndexColorAndImage(indexName: indexName,isIndexOn: isIndexOn ,pathIndex: indexPath.row)
-        cell.addSubview(internalIndexView)
-        internalIndexView.snp.makeConstraints{
-            $0.top.equalToSuperview()
-            $0.width.height.equalTo(30)
-        }
+//        let indexName = self.indexArray[self.internalIndex]
+//        let isIndexOn = calculateInternalIndexCount(indexName: indexName)
+//        let internalIndexView = findInternalIndexColorAndImage(indexName: indexName,isIndexOn: isIndexOn ,pathIndex: indexPath.row)
+//        cell.addSubview(internalIndexView)
+//        internalIndexView.snp.makeConstraints{
+//            $0.top.equalToSuperview()
+//            $0.width.height.equalTo(30)
+//        }
         return cell
     }
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
@@ -558,6 +558,7 @@ extension LocationWeatherIndexView:  UICollectionViewDelegate, UICollectionViewD
         //TODO: 셀에 이미지 클릭하고 호출할 이벤트 넣을 메서드
     }
 }
+
 class CollectionViewRightAlignFlowLayout: UICollectionViewFlowLayout {
     let cellSpacing: CGFloat = 8
     let cellItemSize: Double
